move executeTest method to ContextStateCollector
