+++ failgood/jvm/src/failgood/internal/execution/context/ContextVisitor.kt
@@ -49,7 +49,14 @@ internal class ContextVisitor<GivenType>(
             // we did not yet run a test, so we are going to run this test ourselves
             ranATest = true
 
-            executeTest(testDescription, function, resourcesCloser, isolation, given)
+            contextStateCollector.executeTest(
+                testDescription,
+                function,
+                resourcesCloser,
+                isolation,
+                given,
+                rootContextStartTime
+            )
         } else {
             val resourcesCloser = OnlyResourcesCloser(staticConfig.scope)
             val deferred = staticConfig.scope.async(start = staticConfig.coroutineStart) {
@@ -74,67 +81,6 @@ internal class ContextVisitor<GivenType>(
             }
             contextStateCollector.deferredTestResults[testDescription] = deferred
         }
-    }
-
-    private fun executeTest(
-        testDescription: TestDescription,
-        function: TestLambda<GivenType>,
-        resourcesCloser: ResourcesCloser,
-        isolation: Boolean,
-        given: suspend () -> GivenType
-    ) {
-        contextStateCollector.deferredTestResults[testDescription] =
-            staticConfig.scope.async(start = staticConfig.coroutineStart) {
-
-                val listener = staticConfig.listener
-                listener.testStarted(testDescription)
-                val testResult = try {
-                    withTimeout(staticConfig.timeoutMillis) {
-                        val testContext = TestContext(resourcesCloser, listener, testDescription)
-                        try {
-                            testContext.function(given.invoke())
-                        } catch (e: Throwable) {
-                            val failure = Failure(e)
-                            try {
-                                resourcesCloser.callAfterEach(testContext, failure)
-                            } catch (_: Throwable) {
-                            }
-                            if (isolation) try {
-                                resourcesCloser.closeAutoCloseables()
-                            } catch (_: Throwable) {
-                            }
-                            return@withTimeout failure
-                        }
-                        // test was successful
-                        val success = Success((System.nanoTime() - rootContextStartTime) / 1000)
-                        try {
-                            resourcesCloser.callAfterEach(testContext, success)
-                        } catch (e: Throwable) {
-                            if (isolation) {
-                                try {
-                                    resourcesCloser.closeAutoCloseables()
-                                } catch (e: Throwable) {
-                                    return@withTimeout Failure(e)
-                                }
-                            }
-                            return@withTimeout Failure(e)
-                        }
-                        if (isolation) {
-                            try {
-                                resourcesCloser.closeAutoCloseables()
-                            } catch (e: Throwable) {
-                                return@withTimeout Failure(e)
-                            }
-                        }
-                        success
-                    }
-                } catch (e: TimeoutCancellationException) {
-                    Failure(e)
-                }
-                val testPlusResult = TestPlusResult(testDescription, testResult)
-                listener.testFinished(testPlusResult)
-                testPlusResult
-            }
     }
 
     override suspend fun <ContextDependency> describe(
